Reduce factory demands on FilteItemDict.
By passing it a ready-to-use fs client, we're not relying on the value of of FileItem.file_url to construct one.

Specifically, if the fsspec client requires instantiation with some parameters in the url netloc field, those parameters could be missing from FileItem.file_url.

Generally, FileItem.file_url  has conflicting uses.